Create ranking.py
The code pre-processes the given corpus into a normalized form which involves removing
stopwords and writing each document as a separate line into a new file which serves as 
the input for other functions.

The user inputs a given query which is used for tf-idf calculation with respect to the 
given corpus. The result is stored in a dictionary with the key as the index of each 
document stored in the normalized file. The values are then sorted and the top 10 are
printed according to the key values and their corresponding tf-idf values.

Also,terms like four,FOUR,SIX,six,6,4 are associated with boundary for tf-idf 
computation and ranking.